backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and config
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0

# AI and LLM
openai==1.3.5
langchain==0.0.350
langchain-openai==0.0.2

# Resume parsing and NLP
PyMuPDF==1.23.8
python-docx==0.8.11
pdfplumber==0.9.0
spacy==3.7.2
nltk==3.8.1

# Text processing
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Utilities
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS
python-cors==1.0.0








